add check for pending jobs / subgroups before removing a computer / package / group

+++ lib/CoreLogic.php
@@ -90,7 +90,18 @@ class CoreLogic {
 		return true;
 	}
 	public function removeComputer($id) {
+		$jobs = $this->db->getPendingJobsForComputerDetailPage($id);
+		if(count($jobs) > 0) throw new Exception(LANG['remove_failed_active_jobs']);
+
 		$result = $this->db->removeComputer($id);
+		if(!$result) throw new Exception(LANG['not_found']);
+		return $result;
+	}
+	public function removeComputerGroup($id) {
+		$subgroups = $this->db->getAllComputerGroup($id);
+		if(count($subgroups) > 0) throw new Exception(LANG['remove_failed_subgroups']);
+
+		$result = $this->db->removeComputerGroup($id);
 		if(!$result) throw new Exception(LANG['not_found']);
 		return $result;
 	}
@@ -163,10 +174,23 @@ class CoreLogic {
 	public function removePackage($id) {
 		$package = $this->db->getPackage($id);
 		if(empty($package)) throw new Exception(LANG['not_found']);
+
+		$jobs = $this->db->getPendingJobsForPackageDetailPage($id);
+		if(count($jobs) > 0) throw new Exception(LANG['remove_failed_active_jobs']);
+
 		$path = $package->getFilePath();
 		if(!empty($path)) unlink($path);
+
 		$result = $this->db->removePackage($package->id);
 		if(!$result) throw new Exception(LANG['unknown_error']);
+		return $result;
+	}
+	public function removePackageGroup($id) {
+		$subgroups = $this->db->getAllPackageGroup($id);
+		if(count($subgroups) > 0) throw new Exception(LANG['remove_failed_subgroups']);
+
+		$result = $this->db->removePackageGroup($id);
+		if(!$result) throw new Exception(LANG['not_found']);
 		return $result;
 	}
